Move functions, add README
