+++ packages/nouns-webapp/functions/V0/is-noun-owner/index.ts
@@ -1,5 +1,5 @@
 import { Handler } from '@netlify/functions';
-import { isNounOwner, normalizeNouns, nounsQuery, ownerFilterFactory } from '../theGraph';
+import { isNounOwner, normalizeNouns, nounsQuery } from '../../theGraph';
 
 const handler: Handler = async (event, context) => {
   const graphResponse = await nounsQuery();
