+++ packages/nouns-webapp/functions/V0/noun-owners/index.ts
@@ -1,5 +1,5 @@
 import { Handler } from '@netlify/functions';
-import { normalizeNouns, nounsQuery } from '../theGraph';
+import { normalizeNouns, nounsQuery } from '../../theGraph';
 
 const handler: Handler = async (event, context) => {
   const graphResponse = await nounsQuery();
